Adjust the directory tree
